Tidy up some types and factor out the request body type combo box
