+++ frontend/src/pages/RequestorPage/reducer/index.ts
@@ -1,2 +1,3 @@
 export { useRequestor } from "./reducer";
 export type { ResponsePanelTab, RequestsPanelTab } from "./tabs";
+export type { RequestBodyType, RequestorBody, RequestorState } from "./state";
